pyproject.toml
[tool.poetry]
name = "proyect-etl"
version = "0.1.0"
description = "Proyect of ETL"
authors = ["SantiagoLinux <santiago.gomez.castro.04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
jupyter = "^1.0.0"
pandas = "2.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
matplotlib = "^3.9.2"
apache-airflow = "^2.10.2"
seaborn = "^0.13.2"
kafka-python = "^2.0.2"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
